+++ storage/table/fsm/query.go
@@ -13,6 +13,8 @@ import (
 	sm "github.com/lni/dragonboat/v4/statemachine"
 )
 
+const maxRangeSize uint64 = (4 * 1024 * 1024) - 1024 // 4MiB - 1KiB sentinel.
+
 func commandSnapshot(reader pebble.Reader, tableName string, w io.Writer, stopc <-chan struct{}) (uint64, error) {
 	iter := reader.NewIter(nil)
 	defer iter.Close()
@@ -99,20 +101,19 @@ func rangeLookup(reader pebble.Reader, req *proto.RequestOp_Range) (*proto.Respo
 	defer func() {
 		_ = iter.Close()
 	}()
+	fill, sf := iterFuncsFromReq(req)
+	return iterate(iter, int(req.Limit), fill, sf)
+}
 
-	fill := addKVPair
-	if req.KeysOnly {
-		fill = addKeyOnly
-	} else if req.CountOnly {
-		fill = addCountOnly
+func iterFuncsFromReq(req *proto.RequestOp_Range) (fillEntriesFunc, sizeEntriesFunc) {
+	switch {
+	case req.KeysOnly:
+		return addKeyOnly, sizeKeyOnly
+	case req.CountOnly:
+		return addCountOnly, sizeCountOnly
+	default:
+		return addKVPair, sizeKVPair
 	}
-
-	response := &proto.ResponseOp_Range{}
-	if err = iterate(iter, int(req.Limit), fill, response); err != nil {
-		return nil, err
-	}
-
-	return response, nil
 }
 
 func singleLookup(reader pebble.Reader, req *proto.RequestOp_Range) (*proto.ResponseOp_Range, error) {
@@ -154,24 +155,28 @@ func singleLookup(reader pebble.Reader, req *proto.RequestOp_Range) (*proto.Resp
 // fillEntriesFunc fills proto.RangeResponse response.
 type fillEntriesFunc func(key, value []byte, response *proto.ResponseOp_Range)
 
+// sizeEntriesFunc estimates entry size.
+type sizeEntriesFunc func(key, value []byte) uint64
+
 // iterate until the provided pebble.Iterator is no longer valid or the limit is reached.
 // Apply a function on the key/value pair in every iteration filling proto.RangeResponse.
-func iterate(iter *pebble.Iterator, limit int, f fillEntriesFunc, response *proto.ResponseOp_Range) error {
+func iterate(iter *pebble.Iterator, limit int, f fillEntriesFunc, s sizeEntriesFunc) (*proto.ResponseOp_Range, error) {
+	response := &proto.ResponseOp_Range{}
 	i := 0
 	for iter.First(); iter.Valid(); iter.Next() {
 		k, err := key.DecodeBytes(iter.Key())
 		if err != nil {
-			return err
+			return nil, err
 		}
 
-		if i == limit && limit != 0 {
+		if i == limit && limit != 0 || (uint64(response.SizeVT())+s(k.Key, iter.Value())) >= maxRangeSize {
 			response.More = iter.Next()
 			break
 		}
 		i++
 		f(k.Key, iter.Value(), response)
 	}
-	return nil
+	return response, nil
 }
 
 // addKVPair adds a key/value pair from the provided iterator to the proto.RangeResponse.
@@ -183,6 +188,11 @@ func addKVPair(key, value []byte, response *proto.ResponseOp_Range) {
 	response.Count++
 }
 
+// sizeKVPair takes the full pair size into consideration.
+func sizeKVPair(key, value []byte) uint64 {
+	return uint64(len(key) + len(value))
+}
+
 // addKeyOnly adds a key from the provided iterator to the proto.RangeResponse.
 func addKeyOnly(key, _ []byte, response *proto.ResponseOp_Range) {
 	kv := &proto.KeyValue{Key: make([]byte, len(key))}
@@ -191,9 +201,19 @@ func addKeyOnly(key, _ []byte, response *proto.ResponseOp_Range) {
 	response.Count++
 }
 
+// sizeKeyOnly takes only the key into consideration.
+func sizeKeyOnly(key, _ []byte) uint64 {
+	return uint64(len(key))
+}
+
 // addCountOnly increments number of keys from the provided iterator to the proto.RangeResponse.
 func addCountOnly(_, _ []byte, response *proto.ResponseOp_Range) {
 	response.Count++
+}
+
+// sizeCountOnly for count the size remains constant.
+func sizeCountOnly(_, _ []byte) uint64 {
+	return uint64(0)
 }
 
 // SnapshotRequest to write Command snapshot into provided writer.